requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
black==24.3.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-appconf==1.0.6
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-htmx==1.17.2
django-ninja==1.1.0
django-ninja-extra==0.20.4
googletrans==3.0.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==1.0.4
httpx==0.27.0
huey==2.5.0
hyperframe==5.2.0
idna==2.10
injector==0.21.0
joblib==1.3.2
langdetect==1.0.9
langid==1.1.6
lxml==5.1.0
MarkupSafe==2.1.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
overrides==7.7.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
pyOpenSSL==24.0.0
PyPDF2==3.0.1
pysummarization==1.1.9
python-decouple==3.8
python-docx==1.1.0
python-magic-bin==0.4.14
python3-openid==3.2.0
redis==5.0.3
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
textblob==0.18.0.post0
toolz==0.12.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1